LayoutTests/fast/css/parsing-css-attribute-case-insensitive-value-3-expected.txt

Test the parsing of attribute selector with the case-insensitive value flag (e.g. [foo=bar i]). This test covers invalid selectors of the form [attr=value flags].

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Invalid selectors.
PASS document.querySelector("[foo i]") threw exception SyntaxError: '[foo i]' is not a valid selector..
PASS document.querySelector("[|foo i]") threw exception SyntaxError: '[|foo i]' is not a valid selector..
PASS document.querySelector("[*|foo i]") threw exception SyntaxError: '[*|foo i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo i]") threw exception SyntaxError: '[CyberKit|foo i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo I]") threw exception SyntaxError: '[foo I]' is not a valid selector..
PASS document.querySelector("[|foo I]") threw exception SyntaxError: '[|foo I]' is not a valid selector..
PASS document.querySelector("[*|foo I]") threw exception SyntaxError: '[*|foo I]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo I]") threw exception SyntaxError: '[CyberKit|foo I]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar \"i\"]") threw exception SyntaxError: '[foo=bar "i"]' is not a valid selector..
PASS document.querySelector("[|foo=bar \"i\"]") threw exception SyntaxError: '[|foo=bar "i"]' is not a valid selector..
PASS document.querySelector("[*|foo=bar \"i\"]") threw exception SyntaxError: '[*|foo=bar "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar \"i\"]") threw exception SyntaxError: '[CyberKit|foo=bar "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" \"i\"]") threw exception SyntaxError: '[foo="bar" "i"]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" \"i\"]") threw exception SyntaxError: '[|foo="bar" "i"]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" \"i\"]") threw exception SyntaxError: '[*|foo="bar" "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" \"i\"]") threw exception SyntaxError: '[CyberKit|foo="bar" "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar \"i\"]") threw exception SyntaxError: '[foo~=bar "i"]' is not a valid selector..
PASS document.querySelector("[|foo~=bar \"i\"]") threw exception SyntaxError: '[|foo~=bar "i"]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar \"i\"]") threw exception SyntaxError: '[*|foo~=bar "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar \"i\"]") threw exception SyntaxError: '[CyberKit|foo~=bar "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" \"i\"]") threw exception SyntaxError: '[foo~="bar" "i"]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" \"i\"]") threw exception SyntaxError: '[|foo~="bar" "i"]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" \"i\"]") threw exception SyntaxError: '[*|foo~="bar" "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" \"i\"]") threw exception SyntaxError: '[CyberKit|foo~="bar" "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar \"i\"]") threw exception SyntaxError: '[foo|=bar "i"]' is not a valid selector..
PASS document.querySelector("[|foo|=bar \"i\"]") threw exception SyntaxError: '[|foo|=bar "i"]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar \"i\"]") threw exception SyntaxError: '[*|foo|=bar "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar \"i\"]") threw exception SyntaxError: '[CyberKit|foo|=bar "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" \"i\"]") threw exception SyntaxError: '[foo|="bar" "i"]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" \"i\"]") threw exception SyntaxError: '[|foo|="bar" "i"]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" \"i\"]") threw exception SyntaxError: '[*|foo|="bar" "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" \"i\"]") threw exception SyntaxError: '[CyberKit|foo|="bar" "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar \"i\"]") threw exception SyntaxError: '[foo^=bar "i"]' is not a valid selector..
PASS document.querySelector("[|foo^=bar \"i\"]") threw exception SyntaxError: '[|foo^=bar "i"]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar \"i\"]") threw exception SyntaxError: '[*|foo^=bar "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar \"i\"]") threw exception SyntaxError: '[CyberKit|foo^=bar "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" \"i\"]") threw exception SyntaxError: '[foo^="bar" "i"]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" \"i\"]") threw exception SyntaxError: '[|foo^="bar" "i"]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" \"i\"]") threw exception SyntaxError: '[*|foo^="bar" "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" \"i\"]") threw exception SyntaxError: '[CyberKit|foo^="bar" "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar \"i\"]") threw exception SyntaxError: '[foo$=bar "i"]' is not a valid selector..
PASS document.querySelector("[|foo$=bar \"i\"]") threw exception SyntaxError: '[|foo$=bar "i"]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar \"i\"]") threw exception SyntaxError: '[*|foo$=bar "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar \"i\"]") threw exception SyntaxError: '[CyberKit|foo$=bar "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" \"i\"]") threw exception SyntaxError: '[foo$="bar" "i"]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" \"i\"]") threw exception SyntaxError: '[|foo$="bar" "i"]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" \"i\"]") threw exception SyntaxError: '[*|foo$="bar" "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" \"i\"]") threw exception SyntaxError: '[CyberKit|foo$="bar" "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar \"i\"]") threw exception SyntaxError: '[foo*=bar "i"]' is not a valid selector..
PASS document.querySelector("[|foo*=bar \"i\"]") threw exception SyntaxError: '[|foo*=bar "i"]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar \"i\"]") threw exception SyntaxError: '[*|foo*=bar "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar \"i\"]") threw exception SyntaxError: '[CyberKit|foo*=bar "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" \"i\"]") threw exception SyntaxError: '[foo*="bar" "i"]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" \"i\"]") threw exception SyntaxError: '[|foo*="bar" "i"]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" \"i\"]") threw exception SyntaxError: '[*|foo*="bar" "i"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" \"i\"]") threw exception SyntaxError: '[CyberKit|foo*="bar" "i"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar invalid]") threw exception SyntaxError: '[foo=bar invalid]' is not a valid selector..
PASS document.querySelector("[|foo=bar invalid]") threw exception SyntaxError: '[|foo=bar invalid]' is not a valid selector..
PASS document.querySelector("[*|foo=bar invalid]") threw exception SyntaxError: '[*|foo=bar invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar invalid]") threw exception SyntaxError: '[CyberKit|foo=bar invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" invalid]") threw exception SyntaxError: '[foo="bar" invalid]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" invalid]") threw exception SyntaxError: '[|foo="bar" invalid]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" invalid]") threw exception SyntaxError: '[*|foo="bar" invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" invalid]") threw exception SyntaxError: '[CyberKit|foo="bar" invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar invalid]") threw exception SyntaxError: '[foo~=bar invalid]' is not a valid selector..
PASS document.querySelector("[|foo~=bar invalid]") threw exception SyntaxError: '[|foo~=bar invalid]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar invalid]") threw exception SyntaxError: '[*|foo~=bar invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar invalid]") threw exception SyntaxError: '[CyberKit|foo~=bar invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" invalid]") threw exception SyntaxError: '[foo~="bar" invalid]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" invalid]") threw exception SyntaxError: '[|foo~="bar" invalid]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" invalid]") threw exception SyntaxError: '[*|foo~="bar" invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" invalid]") threw exception SyntaxError: '[CyberKit|foo~="bar" invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar invalid]") threw exception SyntaxError: '[foo|=bar invalid]' is not a valid selector..
PASS document.querySelector("[|foo|=bar invalid]") threw exception SyntaxError: '[|foo|=bar invalid]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar invalid]") threw exception SyntaxError: '[*|foo|=bar invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar invalid]") threw exception SyntaxError: '[CyberKit|foo|=bar invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" invalid]") threw exception SyntaxError: '[foo|="bar" invalid]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" invalid]") threw exception SyntaxError: '[|foo|="bar" invalid]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" invalid]") threw exception SyntaxError: '[*|foo|="bar" invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" invalid]") threw exception SyntaxError: '[CyberKit|foo|="bar" invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar invalid]") threw exception SyntaxError: '[foo^=bar invalid]' is not a valid selector..
PASS document.querySelector("[|foo^=bar invalid]") threw exception SyntaxError: '[|foo^=bar invalid]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar invalid]") threw exception SyntaxError: '[*|foo^=bar invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar invalid]") threw exception SyntaxError: '[CyberKit|foo^=bar invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" invalid]") threw exception SyntaxError: '[foo^="bar" invalid]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" invalid]") threw exception SyntaxError: '[|foo^="bar" invalid]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" invalid]") threw exception SyntaxError: '[*|foo^="bar" invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" invalid]") threw exception SyntaxError: '[CyberKit|foo^="bar" invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar invalid]") threw exception SyntaxError: '[foo$=bar invalid]' is not a valid selector..
PASS document.querySelector("[|foo$=bar invalid]") threw exception SyntaxError: '[|foo$=bar invalid]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar invalid]") threw exception SyntaxError: '[*|foo$=bar invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar invalid]") threw exception SyntaxError: '[CyberKit|foo$=bar invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" invalid]") threw exception SyntaxError: '[foo$="bar" invalid]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" invalid]") threw exception SyntaxError: '[|foo$="bar" invalid]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" invalid]") threw exception SyntaxError: '[*|foo$="bar" invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" invalid]") threw exception SyntaxError: '[CyberKit|foo$="bar" invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar invalid]") threw exception SyntaxError: '[foo*=bar invalid]' is not a valid selector..
PASS document.querySelector("[|foo*=bar invalid]") threw exception SyntaxError: '[|foo*=bar invalid]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar invalid]") threw exception SyntaxError: '[*|foo*=bar invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar invalid]") threw exception SyntaxError: '[CyberKit|foo*=bar invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" invalid]") threw exception SyntaxError: '[foo*="bar" invalid]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" invalid]") threw exception SyntaxError: '[|foo*="bar" invalid]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" invalid]") threw exception SyntaxError: '[*|foo*="bar" invalid]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" invalid]") threw exception SyntaxError: '[CyberKit|foo*="bar" invalid]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar i i]") threw exception SyntaxError: '[foo=bar i i]' is not a valid selector..
PASS document.querySelector("[|foo=bar i i]") threw exception SyntaxError: '[|foo=bar i i]' is not a valid selector..
PASS document.querySelector("[*|foo=bar i i]") threw exception SyntaxError: '[*|foo=bar i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar i i]") threw exception SyntaxError: '[CyberKit|foo=bar i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" i i]") threw exception SyntaxError: '[foo="bar" i i]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" i i]") threw exception SyntaxError: '[|foo="bar" i i]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" i i]") threw exception SyntaxError: '[*|foo="bar" i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" i i]") threw exception SyntaxError: '[CyberKit|foo="bar" i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar i i]") threw exception SyntaxError: '[foo~=bar i i]' is not a valid selector..
PASS document.querySelector("[|foo~=bar i i]") threw exception SyntaxError: '[|foo~=bar i i]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar i i]") threw exception SyntaxError: '[*|foo~=bar i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar i i]") threw exception SyntaxError: '[CyberKit|foo~=bar i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" i i]") threw exception SyntaxError: '[foo~="bar" i i]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" i i]") threw exception SyntaxError: '[|foo~="bar" i i]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" i i]") threw exception SyntaxError: '[*|foo~="bar" i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" i i]") threw exception SyntaxError: '[CyberKit|foo~="bar" i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar i i]") threw exception SyntaxError: '[foo|=bar i i]' is not a valid selector..
PASS document.querySelector("[|foo|=bar i i]") threw exception SyntaxError: '[|foo|=bar i i]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar i i]") threw exception SyntaxError: '[*|foo|=bar i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar i i]") threw exception SyntaxError: '[CyberKit|foo|=bar i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" i i]") threw exception SyntaxError: '[foo|="bar" i i]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" i i]") threw exception SyntaxError: '[|foo|="bar" i i]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" i i]") threw exception SyntaxError: '[*|foo|="bar" i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" i i]") threw exception SyntaxError: '[CyberKit|foo|="bar" i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar i i]") threw exception SyntaxError: '[foo^=bar i i]' is not a valid selector..
PASS document.querySelector("[|foo^=bar i i]") threw exception SyntaxError: '[|foo^=bar i i]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar i i]") threw exception SyntaxError: '[*|foo^=bar i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar i i]") threw exception SyntaxError: '[CyberKit|foo^=bar i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" i i]") threw exception SyntaxError: '[foo^="bar" i i]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" i i]") threw exception SyntaxError: '[|foo^="bar" i i]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" i i]") threw exception SyntaxError: '[*|foo^="bar" i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" i i]") threw exception SyntaxError: '[CyberKit|foo^="bar" i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar i i]") threw exception SyntaxError: '[foo$=bar i i]' is not a valid selector..
PASS document.querySelector("[|foo$=bar i i]") threw exception SyntaxError: '[|foo$=bar i i]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar i i]") threw exception SyntaxError: '[*|foo$=bar i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar i i]") threw exception SyntaxError: '[CyberKit|foo$=bar i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" i i]") threw exception SyntaxError: '[foo$="bar" i i]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" i i]") threw exception SyntaxError: '[|foo$="bar" i i]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" i i]") threw exception SyntaxError: '[*|foo$="bar" i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" i i]") threw exception SyntaxError: '[CyberKit|foo$="bar" i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar i i]") threw exception SyntaxError: '[foo*=bar i i]' is not a valid selector..
PASS document.querySelector("[|foo*=bar i i]") threw exception SyntaxError: '[|foo*=bar i i]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar i i]") threw exception SyntaxError: '[*|foo*=bar i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar i i]") threw exception SyntaxError: '[CyberKit|foo*=bar i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" i i]") threw exception SyntaxError: '[foo*="bar" i i]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" i i]") threw exception SyntaxError: '[|foo*="bar" i i]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" i i]") threw exception SyntaxError: '[*|foo*="bar" i i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" i i]") threw exception SyntaxError: '[CyberKit|foo*="bar" i i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar i first-child]") threw exception SyntaxError: '[foo=bar i first-child]' is not a valid selector..
PASS document.querySelector("[|foo=bar i first-child]") threw exception SyntaxError: '[|foo=bar i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo=bar i first-child]") threw exception SyntaxError: '[*|foo=bar i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar i first-child]") threw exception SyntaxError: '[CyberKit|foo=bar i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" i first-child]") threw exception SyntaxError: '[foo="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" i first-child]") threw exception SyntaxError: '[|foo="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" i first-child]") threw exception SyntaxError: '[*|foo="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" i first-child]") threw exception SyntaxError: '[CyberKit|foo="bar" i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar i first-child]") threw exception SyntaxError: '[foo~=bar i first-child]' is not a valid selector..
PASS document.querySelector("[|foo~=bar i first-child]") threw exception SyntaxError: '[|foo~=bar i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar i first-child]") threw exception SyntaxError: '[*|foo~=bar i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar i first-child]") threw exception SyntaxError: '[CyberKit|foo~=bar i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" i first-child]") threw exception SyntaxError: '[foo~="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" i first-child]") threw exception SyntaxError: '[|foo~="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" i first-child]") threw exception SyntaxError: '[*|foo~="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" i first-child]") threw exception SyntaxError: '[CyberKit|foo~="bar" i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar i first-child]") threw exception SyntaxError: '[foo|=bar i first-child]' is not a valid selector..
PASS document.querySelector("[|foo|=bar i first-child]") threw exception SyntaxError: '[|foo|=bar i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar i first-child]") threw exception SyntaxError: '[*|foo|=bar i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar i first-child]") threw exception SyntaxError: '[CyberKit|foo|=bar i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" i first-child]") threw exception SyntaxError: '[foo|="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" i first-child]") threw exception SyntaxError: '[|foo|="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" i first-child]") threw exception SyntaxError: '[*|foo|="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" i first-child]") threw exception SyntaxError: '[CyberKit|foo|="bar" i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar i first-child]") threw exception SyntaxError: '[foo^=bar i first-child]' is not a valid selector..
PASS document.querySelector("[|foo^=bar i first-child]") threw exception SyntaxError: '[|foo^=bar i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar i first-child]") threw exception SyntaxError: '[*|foo^=bar i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar i first-child]") threw exception SyntaxError: '[CyberKit|foo^=bar i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" i first-child]") threw exception SyntaxError: '[foo^="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" i first-child]") threw exception SyntaxError: '[|foo^="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" i first-child]") threw exception SyntaxError: '[*|foo^="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" i first-child]") threw exception SyntaxError: '[CyberKit|foo^="bar" i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar i first-child]") threw exception SyntaxError: '[foo$=bar i first-child]' is not a valid selector..
PASS document.querySelector("[|foo$=bar i first-child]") threw exception SyntaxError: '[|foo$=bar i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar i first-child]") threw exception SyntaxError: '[*|foo$=bar i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar i first-child]") threw exception SyntaxError: '[CyberKit|foo$=bar i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" i first-child]") threw exception SyntaxError: '[foo$="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" i first-child]") threw exception SyntaxError: '[|foo$="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" i first-child]") threw exception SyntaxError: '[*|foo$="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" i first-child]") threw exception SyntaxError: '[CyberKit|foo$="bar" i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar i first-child]") threw exception SyntaxError: '[foo*=bar i first-child]' is not a valid selector..
PASS document.querySelector("[|foo*=bar i first-child]") threw exception SyntaxError: '[|foo*=bar i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar i first-child]") threw exception SyntaxError: '[*|foo*=bar i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar i first-child]") threw exception SyntaxError: '[CyberKit|foo*=bar i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" i first-child]") threw exception SyntaxError: '[foo*="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" i first-child]") threw exception SyntaxError: '[|foo*="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" i first-child]") threw exception SyntaxError: '[*|foo*="bar" i first-child]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" i first-child]") threw exception SyntaxError: '[CyberKit|foo*="bar" i first-child]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar i \"empty\"]") threw exception SyntaxError: '[foo=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo=bar i \"empty\"]") threw exception SyntaxError: '[|foo=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo=bar i \"empty\"]") threw exception SyntaxError: '[*|foo=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo=bar i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" i \"empty\"]") threw exception SyntaxError: '[foo="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" i \"empty\"]") threw exception SyntaxError: '[|foo="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" i \"empty\"]") threw exception SyntaxError: '[*|foo="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo="bar" i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar i \"empty\"]") threw exception SyntaxError: '[foo~=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo~=bar i \"empty\"]") threw exception SyntaxError: '[|foo~=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar i \"empty\"]") threw exception SyntaxError: '[*|foo~=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo~=bar i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" i \"empty\"]") threw exception SyntaxError: '[foo~="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" i \"empty\"]") threw exception SyntaxError: '[|foo~="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" i \"empty\"]") threw exception SyntaxError: '[*|foo~="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo~="bar" i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar i \"empty\"]") threw exception SyntaxError: '[foo|=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo|=bar i \"empty\"]") threw exception SyntaxError: '[|foo|=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar i \"empty\"]") threw exception SyntaxError: '[*|foo|=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo|=bar i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" i \"empty\"]") threw exception SyntaxError: '[foo|="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" i \"empty\"]") threw exception SyntaxError: '[|foo|="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" i \"empty\"]") threw exception SyntaxError: '[*|foo|="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo|="bar" i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar i \"empty\"]") threw exception SyntaxError: '[foo^=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo^=bar i \"empty\"]") threw exception SyntaxError: '[|foo^=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar i \"empty\"]") threw exception SyntaxError: '[*|foo^=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo^=bar i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" i \"empty\"]") threw exception SyntaxError: '[foo^="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" i \"empty\"]") threw exception SyntaxError: '[|foo^="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" i \"empty\"]") threw exception SyntaxError: '[*|foo^="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo^="bar" i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar i \"empty\"]") threw exception SyntaxError: '[foo$=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo$=bar i \"empty\"]") threw exception SyntaxError: '[|foo$=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar i \"empty\"]") threw exception SyntaxError: '[*|foo$=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo$=bar i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" i \"empty\"]") threw exception SyntaxError: '[foo$="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" i \"empty\"]") threw exception SyntaxError: '[|foo$="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" i \"empty\"]") threw exception SyntaxError: '[*|foo$="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo$="bar" i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar i \"empty\"]") threw exception SyntaxError: '[foo*=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo*=bar i \"empty\"]") threw exception SyntaxError: '[|foo*=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar i \"empty\"]") threw exception SyntaxError: '[*|foo*=bar i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo*=bar i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" i \"empty\"]") threw exception SyntaxError: '[foo*="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" i \"empty\"]") threw exception SyntaxError: '[|foo*="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" i \"empty\"]") threw exception SyntaxError: '[*|foo*="bar" i "empty"]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" i \"empty\"]") threw exception SyntaxError: '[CyberKit|foo*="bar" i "empty"]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar first-child i]") threw exception SyntaxError: '[foo=bar first-child i]' is not a valid selector..
PASS document.querySelector("[|foo=bar first-child i]") threw exception SyntaxError: '[|foo=bar first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo=bar first-child i]") threw exception SyntaxError: '[*|foo=bar first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar first-child i]") threw exception SyntaxError: '[CyberKit|foo=bar first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" first-child i]") threw exception SyntaxError: '[foo="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" first-child i]") threw exception SyntaxError: '[|foo="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" first-child i]") threw exception SyntaxError: '[*|foo="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" first-child i]") threw exception SyntaxError: '[CyberKit|foo="bar" first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar first-child i]") threw exception SyntaxError: '[foo~=bar first-child i]' is not a valid selector..
PASS document.querySelector("[|foo~=bar first-child i]") threw exception SyntaxError: '[|foo~=bar first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar first-child i]") threw exception SyntaxError: '[*|foo~=bar first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar first-child i]") threw exception SyntaxError: '[CyberKit|foo~=bar first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" first-child i]") threw exception SyntaxError: '[foo~="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" first-child i]") threw exception SyntaxError: '[|foo~="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" first-child i]") threw exception SyntaxError: '[*|foo~="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" first-child i]") threw exception SyntaxError: '[CyberKit|foo~="bar" first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar first-child i]") threw exception SyntaxError: '[foo|=bar first-child i]' is not a valid selector..
PASS document.querySelector("[|foo|=bar first-child i]") threw exception SyntaxError: '[|foo|=bar first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar first-child i]") threw exception SyntaxError: '[*|foo|=bar first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar first-child i]") threw exception SyntaxError: '[CyberKit|foo|=bar first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" first-child i]") threw exception SyntaxError: '[foo|="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" first-child i]") threw exception SyntaxError: '[|foo|="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" first-child i]") threw exception SyntaxError: '[*|foo|="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" first-child i]") threw exception SyntaxError: '[CyberKit|foo|="bar" first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar first-child i]") threw exception SyntaxError: '[foo^=bar first-child i]' is not a valid selector..
PASS document.querySelector("[|foo^=bar first-child i]") threw exception SyntaxError: '[|foo^=bar first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar first-child i]") threw exception SyntaxError: '[*|foo^=bar first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar first-child i]") threw exception SyntaxError: '[CyberKit|foo^=bar first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" first-child i]") threw exception SyntaxError: '[foo^="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" first-child i]") threw exception SyntaxError: '[|foo^="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" first-child i]") threw exception SyntaxError: '[*|foo^="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" first-child i]") threw exception SyntaxError: '[CyberKit|foo^="bar" first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar first-child i]") threw exception SyntaxError: '[foo$=bar first-child i]' is not a valid selector..
PASS document.querySelector("[|foo$=bar first-child i]") threw exception SyntaxError: '[|foo$=bar first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar first-child i]") threw exception SyntaxError: '[*|foo$=bar first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar first-child i]") threw exception SyntaxError: '[CyberKit|foo$=bar first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" first-child i]") threw exception SyntaxError: '[foo$="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" first-child i]") threw exception SyntaxError: '[|foo$="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" first-child i]") threw exception SyntaxError: '[*|foo$="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" first-child i]") threw exception SyntaxError: '[CyberKit|foo$="bar" first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar first-child i]") threw exception SyntaxError: '[foo*=bar first-child i]' is not a valid selector..
PASS document.querySelector("[|foo*=bar first-child i]") threw exception SyntaxError: '[|foo*=bar first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar first-child i]") threw exception SyntaxError: '[*|foo*=bar first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar first-child i]") threw exception SyntaxError: '[CyberKit|foo*=bar first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" first-child i]") threw exception SyntaxError: '[foo*="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" first-child i]") threw exception SyntaxError: '[|foo*="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" first-child i]") threw exception SyntaxError: '[*|foo*="bar" first-child i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" first-child i]") threw exception SyntaxError: '[CyberKit|foo*="bar" first-child i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar \"empty\" i]") threw exception SyntaxError: '[foo=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo=bar \"empty\" i]") threw exception SyntaxError: '[|foo=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo=bar \"empty\" i]") threw exception SyntaxError: '[*|foo=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo=bar "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" \"empty\" i]") threw exception SyntaxError: '[foo="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" \"empty\" i]") threw exception SyntaxError: '[|foo="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" \"empty\" i]") threw exception SyntaxError: '[*|foo="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo="bar" "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar \"empty\" i]") threw exception SyntaxError: '[foo~=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo~=bar \"empty\" i]") threw exception SyntaxError: '[|foo~=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar \"empty\" i]") threw exception SyntaxError: '[*|foo~=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo~=bar "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" \"empty\" i]") threw exception SyntaxError: '[foo~="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" \"empty\" i]") threw exception SyntaxError: '[|foo~="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" \"empty\" i]") threw exception SyntaxError: '[*|foo~="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo~="bar" "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar \"empty\" i]") threw exception SyntaxError: '[foo|=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo|=bar \"empty\" i]") threw exception SyntaxError: '[|foo|=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar \"empty\" i]") threw exception SyntaxError: '[*|foo|=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo|=bar "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" \"empty\" i]") threw exception SyntaxError: '[foo|="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" \"empty\" i]") threw exception SyntaxError: '[|foo|="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" \"empty\" i]") threw exception SyntaxError: '[*|foo|="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo|="bar" "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar \"empty\" i]") threw exception SyntaxError: '[foo^=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo^=bar \"empty\" i]") threw exception SyntaxError: '[|foo^=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar \"empty\" i]") threw exception SyntaxError: '[*|foo^=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo^=bar "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" \"empty\" i]") threw exception SyntaxError: '[foo^="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" \"empty\" i]") threw exception SyntaxError: '[|foo^="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" \"empty\" i]") threw exception SyntaxError: '[*|foo^="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo^="bar" "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar \"empty\" i]") threw exception SyntaxError: '[foo$=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo$=bar \"empty\" i]") threw exception SyntaxError: '[|foo$=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar \"empty\" i]") threw exception SyntaxError: '[*|foo$=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo$=bar "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" \"empty\" i]") threw exception SyntaxError: '[foo$="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" \"empty\" i]") threw exception SyntaxError: '[|foo$="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" \"empty\" i]") threw exception SyntaxError: '[*|foo$="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo$="bar" "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar \"empty\" i]") threw exception SyntaxError: '[foo*=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo*=bar \"empty\" i]") threw exception SyntaxError: '[|foo*=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar \"empty\" i]") threw exception SyntaxError: '[*|foo*=bar "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo*=bar "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" \"empty\" i]") threw exception SyntaxError: '[foo*="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" \"empty\" i]") threw exception SyntaxError: '[|foo*="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" \"empty\" i]") threw exception SyntaxError: '[*|foo*="bar" "empty" i]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" \"empty\" i]") threw exception SyntaxError: '[CyberKit|foo*="bar" "empty" i]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=bar l]") threw exception SyntaxError: '[foo=bar l]' is not a valid selector..
PASS document.querySelector("[|foo=bar l]") threw exception SyntaxError: '[|foo=bar l]' is not a valid selector..
PASS document.querySelector("[*|foo=bar l]") threw exception SyntaxError: '[*|foo=bar l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=bar l]") threw exception SyntaxError: '[CyberKit|foo=bar l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo=\"bar\" l]") threw exception SyntaxError: '[foo="bar" l]' is not a valid selector..
PASS document.querySelector("[|foo=\"bar\" l]") threw exception SyntaxError: '[|foo="bar" l]' is not a valid selector..
PASS document.querySelector("[*|foo=\"bar\" l]") threw exception SyntaxError: '[*|foo="bar" l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo=\"bar\" l]") threw exception SyntaxError: '[CyberKit|foo="bar" l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=bar l]") threw exception SyntaxError: '[foo~=bar l]' is not a valid selector..
PASS document.querySelector("[|foo~=bar l]") threw exception SyntaxError: '[|foo~=bar l]' is not a valid selector..
PASS document.querySelector("[*|foo~=bar l]") threw exception SyntaxError: '[*|foo~=bar l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=bar l]") threw exception SyntaxError: '[CyberKit|foo~=bar l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo~=\"bar\" l]") threw exception SyntaxError: '[foo~="bar" l]' is not a valid selector..
PASS document.querySelector("[|foo~=\"bar\" l]") threw exception SyntaxError: '[|foo~="bar" l]' is not a valid selector..
PASS document.querySelector("[*|foo~=\"bar\" l]") threw exception SyntaxError: '[*|foo~="bar" l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo~=\"bar\" l]") threw exception SyntaxError: '[CyberKit|foo~="bar" l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=bar l]") threw exception SyntaxError: '[foo|=bar l]' is not a valid selector..
PASS document.querySelector("[|foo|=bar l]") threw exception SyntaxError: '[|foo|=bar l]' is not a valid selector..
PASS document.querySelector("[*|foo|=bar l]") threw exception SyntaxError: '[*|foo|=bar l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=bar l]") threw exception SyntaxError: '[CyberKit|foo|=bar l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo|=\"bar\" l]") threw exception SyntaxError: '[foo|="bar" l]' is not a valid selector..
PASS document.querySelector("[|foo|=\"bar\" l]") threw exception SyntaxError: '[|foo|="bar" l]' is not a valid selector..
PASS document.querySelector("[*|foo|=\"bar\" l]") threw exception SyntaxError: '[*|foo|="bar" l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo|=\"bar\" l]") threw exception SyntaxError: '[CyberKit|foo|="bar" l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=bar l]") threw exception SyntaxError: '[foo^=bar l]' is not a valid selector..
PASS document.querySelector("[|foo^=bar l]") threw exception SyntaxError: '[|foo^=bar l]' is not a valid selector..
PASS document.querySelector("[*|foo^=bar l]") threw exception SyntaxError: '[*|foo^=bar l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=bar l]") threw exception SyntaxError: '[CyberKit|foo^=bar l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo^=\"bar\" l]") threw exception SyntaxError: '[foo^="bar" l]' is not a valid selector..
PASS document.querySelector("[|foo^=\"bar\" l]") threw exception SyntaxError: '[|foo^="bar" l]' is not a valid selector..
PASS document.querySelector("[*|foo^=\"bar\" l]") threw exception SyntaxError: '[*|foo^="bar" l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo^=\"bar\" l]") threw exception SyntaxError: '[CyberKit|foo^="bar" l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=bar l]") threw exception SyntaxError: '[foo$=bar l]' is not a valid selector..
PASS document.querySelector("[|foo$=bar l]") threw exception SyntaxError: '[|foo$=bar l]' is not a valid selector..
PASS document.querySelector("[*|foo$=bar l]") threw exception SyntaxError: '[*|foo$=bar l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=bar l]") threw exception SyntaxError: '[CyberKit|foo$=bar l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo$=\"bar\" l]") threw exception SyntaxError: '[foo$="bar" l]' is not a valid selector..
PASS document.querySelector("[|foo$=\"bar\" l]") threw exception SyntaxError: '[|foo$="bar" l]' is not a valid selector..
PASS document.querySelector("[*|foo$=\"bar\" l]") threw exception SyntaxError: '[*|foo$="bar" l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo$=\"bar\" l]") threw exception SyntaxError: '[CyberKit|foo$="bar" l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=bar l]") threw exception SyntaxError: '[foo*=bar l]' is not a valid selector..
PASS document.querySelector("[|foo*=bar l]") threw exception SyntaxError: '[|foo*=bar l]' is not a valid selector..
PASS document.querySelector("[*|foo*=bar l]") threw exception SyntaxError: '[*|foo*=bar l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=bar l]") threw exception SyntaxError: '[CyberKit|foo*=bar l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector("[foo*=\"bar\" l]") threw exception SyntaxError: '[foo*="bar" l]' is not a valid selector..
PASS document.querySelector("[|foo*=\"bar\" l]") threw exception SyntaxError: '[|foo*="bar" l]' is not a valid selector..
PASS document.querySelector("[*|foo*=\"bar\" l]") threw exception SyntaxError: '[*|foo*="bar" l]' is not a valid selector..
PASS document.querySelector("[CyberKit|foo*=\"bar\" l]") threw exception SyntaxError: '[CyberKit|foo*="bar" l]' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS successfullyParsed is true

TEST COMPLETE
